Work around bug requiring to hit esc,h twice to filter the message
In some cases, <edit-type> command prompts whether to convert content.
This prompt is not configurable, so we press <Return> to either select
default option, or to enter content view. Then, we press <Return again
to be sure we entered content view (if we're already here, nothing
happens unless user has it bound to some action). Then we <exit> from
content view.